Show an example of more modular code
Here is an example of how to organize the autonomous code in a way that
would not require maintaining separate git branches. Effectively you can
setup the code to support different autonomous strategies leveraging the
power of `Commands.sequence` and tying smaller subcommands together.

Feel free to use, or save this for after competitions as an example for
next year.